==> munge/00.02.FixorCalculateOffsets.Rmd <==
---
title: "Fix or Calculate Offsets"
author: "Nicole Barker"
date: "Last run: Jan 8, 2018"
output: 
  word_document:
    reference_docx: ../styles/ReportFormat_1.docx
---

## Script Abstract

## Background


``` {r setup, echo=F, message=F, warning=F}
require(knitr)
opts_knit$set(root.dir = '..')
knitr::opts_chunk$set(
  fig.path = "figures/"
)
```

``` {r load.project, message=F}
require(ProjectTemplate)
load.project()
options(digits=12)
```

### Load pre-processed bird data

* No screening for temporal alignment or distance to edge
* This should be all the bird data received from Trish in November 2017.

``` {r}
birddat1 <- read.csv("data/birddata_preprocess1.csv", header=T)
colnames(birddat1)
colnames(birddat1)[which(colnames(birddat1) == "PCODE_derived")] <- "PCODE"
colnames(birddat1)[which(colnames(birddat1) == "SS_derived")] <- "SS"
```

**Check for Missing Data**

``` {r}
kable(as.data.frame(do.call(rbind,lapply(birddat1, function(x) {sum(is.na(x))})))) # check for missing data
```


### Re-calculate offsets since we're missing a bunch 

#### Simpler method without - just corrects for method

Code supplied by Peter Solymos on Jan 2, 2018.  
 
* Note that this method uses only max distance and max duration, so it only corrects for methodology. Not location- or time-specific covariates.  
 
``` {r, eval=T} 
require(QPAD) 
x <- birddat1 
str(x) 
table(x$MaxDuration, useNA="a") 
table(x$Maxdist, useNA="a") 
levels(x$Maxdist)[levels(x$Maxdist) == "Unlimited"] <- "Inf" 
x$MaxDistance <- as.numeric(as.character(x$Maxdist)) / 100 
table(x$MaxDistance, useNA="a") 
 
# numbers taken from estimates 
phi_tota <- exp(0.2773876) 
tau_tota <- exp(0.8422723) 
x$p <- sra_fun(x$MaxDuration, phi_tota) 
x$A <- ifelse(is.finite(x$MaxDistance), x$MaxDistance^2*pi, tau_tota^2*phi_tota) 
x$q <- ifelse(is.finite(x$MaxDistance), edr_fun(x$MaxDistance, tau_tota), 1) 
summary(x$p) 
summary(x$A) 
summary(x$q) 
 
x$Exposure_simple <- with(x, p * A * q) 
x$Offset_simple <- log(x$Exposure_simple) 
summary(x$Offset_simple) 
length(x$Offset_simple) 
 
offsets_simple <- x[c("PKEY", "Offset_simple", "Exposure_simple")]

sum(is.na(x$Offset_simple)) # how many PKEYs don't have offsets 
``` 


#### With Covariates
Code supplied by Peter Solymos on Jan 7, 2018. 

https://github.com/psolymos/bamanalytics/blob/master/projects/bamdb/offsets_TOTA.R

I should do some branching but I don't have time to figure it out right now. 

``` {r, eval=T}
library(QPAD) # latest estimates
library(maptools) # for sunrise calculations
library(mefa4) # data manipulation

dat <- read.csv("data/birddata_preprocess1.csv", header=T)
dat.qs <- dat

## max duration
dat$MAXDUR <- dat$MaxDuration
table(dat$MAXDUR, useNA="always")

## max distance
dat$MAXDIS <- droplevels(dat$Maxdist)
levels(dat$MAXDIS) <- toupper(levels(dat$MAXDIS))
levels(dat$MAXDIS)[levels(dat$MAXDIS) == "UNLIMITED"] <- "Inf"
dat$MAXDIS <- as.numeric(as.character(dat$MAXDIS)) / 100
table(dat$MAXDIS, useNA="always")

## Date/time components
chr <- paste0(dat$YYYY, "-", dat$MM, "-", dat$DD, " ", dat$HR, ":", dat$MIN)
myDATE <- strptime(chr, format="%Y-%m-%e %H:%M")
dat$DATE <- myDATE
summary(dat$DATE)
class(myDATE)
class(dat$DATE)

dat$JULIAN <- dat$DATE$yday
dat$JDAY <- myDATE$yday / 365
summary(dat$JDAY)
## prevent too far extrapolation
#dat$JDAY[dat$JDAY < 0.35 | dat$JDAY > 0.55] <- NA
hist(dat$JDAY)

Coor <- as.matrix(cbind(as.numeric(dat$X_coor), as.numeric(dat$Y_coor)))
JL <- as.POSIXct(myDATE)
subset <- rowSums(is.na(Coor))==0 & !is.na(JL)
sr <- sunriset(Coor[subset,], JL[subset], direction="sunrise", POSIXct.out=FALSE) * 24
dat$srise <- NA
dat$srise[subset] <- sr
dat$start_time <- dat$DATE$hour + dat$DATE$min/60

dat$MDT_offset <- -6
dat$MDT_offset[dat$TZ == "America/Dawson_Creek"] <- -7
dat$MDT_offset[dat$TZ == "America/Vancouver"] <- -8
dat$MDT_offset <- dat$MDT_offset + 6
table(dat$MDT_offset, useNA="always")

dat$TSSR <- (dat$start_time - dat$srise + dat$MDT_offset) / 24
dat$TSSR_orig <- dat$TSSR # keep a full copy
dat$TSSR[dat$start_time > 12] <- NA # after noon
summary(dat$TSSR)
summary(dat$start_time)
hist(dat$TSSR)

dat$TREE <- dat$tree
summary(dat$TREE)
dat$TREE[dat$TREE > 100] <- NA
dat$TREE[dat$TREE < 0] <- NA
dat$TREE <- dat$TREE / 100
summary(dat$TREE)
hist(dat$TREE)

(ltnalc <- read.csv("data/lookup.nalcms.csv", header=T))
table(dat$NALCMS05, useNA="always")
dat$NALCMS05[dat$NALCMS05 < 0] <- 0
compare_sets(dat$NALCMS05, ltnalc$Value)
dat$LCC2 <- reclass(dat$NALCMS05, ltnalc[,c("Value", "LCC2")], allow_NA=TRUE)
table(dat$NALCMS05, dat$LCC2, useNA="always")
dat$LCC4 <- reclass(dat$NALCMS05, ltnalc[,c("Value", "LCC4")], allow_NA=TRUE)
table(dat$NALCMS05, dat$LCC4, useNA="always")
boxplot(TREE ~ LCC4, dat)

load_BAM_QPAD(3)
SPP <- getBAMspecieslist()
dat$JDAY2 <- dat$JDAY^2
dat$TSSR2 <- dat$TSSR^2

Xp <- cbind("(Intercept)"=1, as.matrix(dat[,c("TSSR","JDAY","TSSR2","JDAY2")]))
Xq <- cbind("(Intercept)"=1, TREE=dat$TREE,
    LCC2OpenWet=ifelse(dat$LCC2=="OpenWet", 1, 0),
    LCC4Conif=ifelse(dat$LCC4=="Conif", 1, 0),
    LCC4Open=ifelse(dat$LCC4=="Open", 1, 0),
    LCC4Wet=ifelse(dat$LCC4=="Wet", 1, 0))
OFF <- matrix(NA, nrow(dat), length(SPP))
rownames(OFF) <- dat$PKEY
colnames(OFF) <- SPP

## load TOTA results
load("cache/TOTA_BAMCOEFS_QPAD_v3.rda")
(mods <- getBAMmodellist())
(sra_mods <- names(mods$sra)[!grepl("DSLS", mods$sra)])
getBAMspecieslist()

spp <- "TOTA"
p <- rep(NA, nrow(dat))
A <- q <- p
## constant for NA cases
(cf0 <- exp(unlist(coefBAMspecies(spp, 0, 0))))
## best model
(mi <- bestmodelBAMspecies(spp, model.sra=sra_mods, type="BIC"))
(cfi <- coefBAMspecies(spp, mi$sra, mi$edr))

Xp2 <- Xp[,names(cfi$sra),drop=FALSE]
OKp <- rowSums(is.na(Xp2)) == 0
Xq2 <- Xq[,names(cfi$edr),drop=FALSE]
OKq <- rowSums(is.na(Xq2)) == 0

p[!OKp] <- sra_fun(dat$MAXDUR[!OKp], cf0[1])
unlim <- ifelse(dat$MAXDIS[!OKq] == Inf, TRUE, FALSE)
A[!OKq] <- ifelse(unlim, pi * cf0[2]^2, pi * dat$MAXDIS[!OKq]^2)
q[!OKq] <- ifelse(unlim, 1, edr_fun(dat$MAXDIS[!OKq], cf0[2]))

phi1 <- exp(drop(Xp2[OKp,,drop=FALSE] %*% cfi$sra))
tau1 <- exp(drop(Xq2[OKq,,drop=FALSE] %*% cfi$edr))
p[OKp] <- sra_fun(dat$MAXDUR[OKp], phi1)
unlim <- ifelse(dat$MAXDIS[OKq] == Inf, TRUE, FALSE)
A[OKq] <- ifelse(unlim, pi * tau1^2, pi * dat$MAXDIS[OKq]^2)
q[OKq] <- ifelse(unlim, 1, edr_fun(dat$MAXDIS[OKq], tau1))

ii <- which(p == 0)
p[ii] <- sra_fun(dat$MAXDUR[ii], cf0[1])

dat$CORRECTION <- data.frame(p=p, A=A, q=q)
summary(dat$CORRECTION)
dat$OFFSET <- rowSums(log(dat$CORRECTION))
summary(dat$OFFSET)
nrow(dat[is.na(dat$OFFSET),]) # methodology is unknown --> drop!

dat.known <- dat[!is.na(dat$OFFSET),]
    
offsets <- offsets_2018.01.08 <- dat.known[c("PKEY", "OFFSET")]

offsets_complex <- dat[c("PKEY", "OFFSET")]
offsets_complex <- offsets_complex[!duplicated(offsets_complex$PKEY),]
```

### combine offsets

``` {r}
offsets_all <- merge(offsets_simple, offsets_complex, by="PKEY", all=T)
cache("offsets_all")
```
