==> munge/01.03.Merge_Full_VRI_Data.Rmd <==
---
title: "Merge Forsite and Canfor TFL VRI files together"
author: "Nicole Barker"
date: "Last run: Jan 8, 2018"
output: 
  word_document:
    reference_docx: ../styles/ReportFormat_1.docx
---

## Script Abstract
Identify which columns are needed from both VRI files (Canfor and Forsite). Add blank columns to Canfor file so it can be combind with the Forsite file. Combine the two files together and modify the factors to have the correct levels. 

## Background

Forsite's VRI file has blank attributes for tree farm licences (TFLs) that Canfor has data for. Canfor gave us their VRI data for those forest stands. I separately quality-checked and processed the two VRI datasets and now it's time to combine them, after doing a bit more processing. 

**FILES**

1. vri_forsite_zSpGroup_Rankfixed.RData
2. vri_canfortfl.RData

In this script, I: 

* Look for overlapping columns between two VRI files
* create blank columns in Canfor table where I need them
* Combine two VRI files together
* Convert date columns to date format 
* Pull year out from date
* Classify stands to a stand age class if not already done
* Classify stands to a stange height class if not already done
* Create multiple files for future use 
    * Forsite VRI: for analyses using Kari/Forsite's 4 derived attributes (will exclude Canfor TFL); and
    * All VRI: for density models (using variables from VRI directly, not derived)


``` {r setup, echo=F, message=F, warning=F}
require(knitr)
opts_knit$set(root.dir = '..')
knitr::opts_chunk$set(
  fig.path = "figures/"
)
```

``` {r load.project, message=F}
require(ProjectTemplate)
load.project()
options(digits=12)
```

### Look for overlapping columns between two VRI files

``` {r}
load("cache/vri_forsite_zSpGroup_Rankfixed.RData")
vri_forsite <- vri_forsite_zSpGroup_Rankfixed; rm(vri_forsite_zSpGroup_Rankfixed)
colnames(vri_forsite)[which(colnames(vri_forsite) == "FinalRank")] <- "Rank"

forsite_colnames <- data.frame(ColName=colnames(vri_forsite), In_ForsiteVRI=as.character("Yes"))


load("cache/vri_canfortfl_mostfields.RData")
vri_canfortfl <- vri_canfortfl_mostfields; rm(vri_canfortfl_mostfields)
canfor_colnames <- data.frame(ColName=colnames(vri_canfortfl), In_vri_canfortfl=as.character("Yes"))
```

``` {r}
all_colnames <- merge(canfor_colnames, forsite_colnames, by="ColName", all=T)

all_colnames$In_vri_canfortfl <- as.character(all_colnames$In_vri_canfortfl)
all_colnames$In_ForsiteVRI <- as.character(all_colnames$In_ForsiteVRI)

all_colnames$In_vri_canfortfl[is.na(all_colnames$In_vri_canfortfl)] <- ""
all_colnames$In_ForsiteVRI[is.na(all_colnames$In_ForsiteVRI)] <- ""

kable(all_colnames, row.names=F)
```

#### Add blank columns for those I need from the Forsite layer (i.e. derived fields)

``` {r}
vri_canfortfl$BecZone <- NA                 
vri_canfortfl$SpeciesGroup <- NA
vri_canfortfl$Age <- NA
vri_canfortfl$Height <- NA
vri_canfortfl$Rank <- NA
vri_canfortfl$FD_PERCENT <- NA
vri_canfortfl$PL_PERCENT <- NA
vri_canfortfl$AT_PERCENT <- NA
vri_canfortfl$Decid_PERCENT <- NA
vri_canfortfl$polygon_area_cal <- NA
vri_canfortfl$FEATURE_ID <- NA
vri_forsite$OBJECTID <- NA

canfor_colnames <- data.frame(ColName=colnames(vri_canfortfl), In_vri_canfortfl=as.character("Yes"))

all_colnames <- merge(canfor_colnames, forsite_colnames, by="ColName", all=T)

all_colnames$In_vri_canfortfl <- as.character(all_colnames$In_vri_canfortfl)
all_colnames$In_ForsiteVRI <- as.character(all_colnames$In_ForsiteVRI)

all_colnames$In_vri_canfortfl[is.na(all_colnames$In_vri_canfortfl)] <- ""
all_colnames$In_ForsiteVRI[is.na(all_colnames$In_ForsiteVRI)] <- ""

kable(all_colnames, row.names=F)
```

## Merge two VRI files together

#### Subset for common set of columns and create a layer + polygon unique identifier that works across spatial files
``` {r}
commoncols <- all_colnames[all_colnames$In_vri_canfortfl == "Yes" & all_colnames$In_ForsiteVRI == "Yes",]$ColName

vri_forsite2 <- vri_forsite[which(colnames(vri_forsite) %in% commoncols)]
vri_forsite2$ID.Layer.Polygon <- paste(vri_forsite2$VRI_Layer, vri_forsite2$FEATURE_ID, sep=".")
nrow(vri_forsite2)
length(unique(vri_forsite2$ID.Layer.Polygon))

vri_canfortfl2 <- vri_canfortfl[which(colnames(vri_canfortfl) %in% commoncols)]
vri_canfortfl2$ID.Layer.Polygon <- paste(vri_canfortfl2$VRI_Layer, vri_canfortfl2$OBJECTID, sep=".")
nrow(vri_canfortfl2)
length(unique(vri_canfortfl2$ID.Layer.Polygon))
```

#### rbind together

``` {r}
vri_both <- rbind(vri_canfortfl2, vri_forsite2)
nrow(vri_both)
length(unique(vri_both$ID.Layer.Polygon))
```

## More pre-processing

### Date pre-processing

``` {r}
# function to extract year from date columns
F.convertdate.extractyear <- function(x){
  dates <- as.Date(unlist(lapply(strsplit(as.character(x), " "), function(y) {y[1]})), format="%m/%d/%Y")
  years <- format(dates,'%Y')
}
```


Compare Reference Year to Reference Date
``` {r}
vri_both$ReferenceYear <- F.convertdate.extractyear(vri_both$REFERENCE_DATE) #take year from reference Date if it's available

refdate_year <- vri_both[c("REFERENCE_DATE", "REFERENCE_YEAR", "ReferenceYear")]
refdate_year <- refdate_year[!is.na(refdate_year$REFERENCE_YEAR),] #remove rows where supplied Reference Year is blank

kable(head(refdate_year[is.na(refdate_year$ReferenceYear),]), row.names=F)
```

Some polygons have reference year even if they don't have reference date.

Now to check if reference date and reference year have the same year if they're both provided. 

``` {r}
refdate_year2 <- refdate_year[!is.na(refdate_year$ReferenceYear),]
refdate_year2$DATE_YEAR_same <- refdate_year2$REFERENCE_YEAR != refdate_year2$ReferenceYear #return true if provided and calculated years are different
sum(refdate_year2$DATE_YEAR_same) # for how many rows was my statement TRUE
```

* If the REFERENCE_DATE is provided, it is the same year as REFERENCE_YEAR.
* Therefore, I can simply use REFERENCE_YEAR instead of calculating it. 

#### Extract various years out of datasets

``` {r, warning=F, message=F}
vri_both$InputYear <- F.convertdate.extractyear(vri_both$INPUT_DATE)
vri_both$ProjectedYear <- F.convertdate.extractyear(vri_both$PROJECTED_DATE)
vri_both$InterpretYear <- F.convertdate.extractyear(vri_both$INTERPRETATION_DATE)
vri_both$ReferenceYear <- vri_both$REFERENCE_YEAR
vri_both$AttributionYear <- F.convertdate.extractyear(vri_both$ATTRIBUTION_BASE_DATE)
vri_both$DistYear_Nonlog <- F.convertdate.extractyear(vri_both$EARLIEST_NONLOGGING_DIST_DATE)
vri_both$DistYear_Log <- F.convertdate.extractyear(vri_both$HARVEST_DATE)
vri_both$DisturbanceYear <- apply(vri_both[c("DistYear_Nonlog", "DistYear_Log")], MARGIN=1, FUN=function(x) {max(x, na.rm=T)}) 
```

### Stand Age classification

``` {r}
ageclassbreaks <- c(0, 3, 30, 80, 120, 250, 1000)
vri_both$AgeClass_calc <- cut(vri_both$PROJ_AGE_1, breaks = ageclassbreaks, include.lowest = TRUE, labels = c("0-2 yrs", "3-30 yrs", "31-80 yrs", "81-120 yrs", "121-250 yrs", ">250 yrs"))

vri_both$AgeClass_calc <- ordered(vri_both$AgeClass_calc) # make an ordered factor

kable(rbind(head(vri_both[c("PROJ_AGE_1", "AgeClass_calc", "VRI_Layer")]),
            tail(vri_both[c("PROJ_AGE_1", "AgeClass_calc", "VRI_Layer")])))

```

##### What about stands without age (e.g., used for other purposes)

``` {r}
head(vri_both[is.na(vri_both$AgeClass_calc), c("PROJ_AGE_1", "AgeClass_calc", "Age", "NON_PRODUCTIVE_DESCRIPTOR_CD")])
vri_both[is.na(vri_both$AgeClass_calc),]$AgeClass_calc <- "0-2 yrs" #replace non-forested stands with 0-2 years
```

### Stand Height classification

``` {r}
heightclassbreaks <- c(0, 10.5, 19.5, 28.5, 100)
vri_both$HeightClass_calc <- cut(vri_both$PROJ_HEIGHT_1, breaks = heightclassbreaks, include.lowest = TRUE, labels = c("<10.5 m", "10.5-19.4 m", "19.5-28.4 m", "> 28.5 m"))

vri_both$HeightClass_calc <- ordered(vri_both$HeightClass_calc) # make an ordered factor

kable(rbind(head(vri_both[c("PROJ_HEIGHT_1", "HeightClass_calc", "VRI_Layer")]),
            tail(vri_both[c("PROJ_HEIGHT_1", "HeightClass_calc", "VRI_Layer")])))
```

##### What about stands without height (e.g., non-forested)

``` {r}
head(vri_both[is.na(vri_both$HeightClass_calc), c("PROJ_HEIGHT_1", "HeightClass_calc", "Height", "NON_PRODUCTIVE_DESCRIPTOR_CD")])
vri_both[is.na(vri_both$HeightClass_calc),]$HeightClass_calc <- "<10.5 m" #replace non-forested stands with < 10.5 m
```

## Quality Checks

``` {r}
nrow(vri_both)
length(unique(vri_both$ID.Layer.Polygon))
colnames(vri_both)
# these quality checks take too long to run

#kable(as.data.frame(do.call(rbind, lapply(vri_both, function(x) {length(unique(x))}))), caption="Number of unique values in each column") 

#kable(as.data.frame(do.call(rbind, lapply(vri_both, function(x) {sum(is.na(x))}))), caption="Number of rows with NA values in each field")
```


## Save files


``` {r, eval=T}
cache("vri_both")

vri_canfortfl_postclean <- vri_both[vri_both$VRI_Layer == "Canfor_TFL",]
cache("vri_canfortfl_postclean")

vri_forsite_postclean <- vri_both[vri_both$VRI_Layer == "Forsite_VRI",]
cache("vri_forsite_postclean")
```
