==> munge/03.01.FinalDataPrep.Rmd <==
---
title: "Final Data Preparation"
author: "Nicole Barker"
date: "Last run: Jan 9, 2018" 
output: 
  word_document:
    reference_docx: ../styles/ReportFormat_1.docx
---

## Script Abstract

## Background
In previous steps: 

* I added offsets into the table
* I filtered dates based on mismatch between VRI data and bird surveys
* I expect to filter based on proximity of point count to stand edge once that data is available. NOT CURRENTLY DONE. 

However, I still need to do a bit more data processing before I can do analyses. 

In this script: 

* Subset to desired behaviours
* I sum across species to develop a total bird count per survey 


**I may also:**

* Explore relationships between project (PCODE) and apparent total bird count


``` {r setup, echo=F, message=F, warning=F}
require(knitr)
opts_knit$set(root.dir = '..')
knitr::opts_chunk$set(
  fig.path = "figures/"
)
```

``` {r load.project, message=F}
require(ProjectTemplate)
load.project()
options(digits=12)
```

### Load intersected bird data, combine with methods, and subset to columns needed for exploration of bird counts per different survey methods

``` {r}
load("cache/vri_bird_tempaligned.RData") #(PKEYs quality-checked for use)
colnames(vri_bird_tempaligned)
#load("cache/vri_bird_tempaligned-distfiltered.RData") # this doesn't exist yet

birddat1 <- read.csv("data/birddata_preprocess1.csv", header=T)
colnames(birddat1)
colnames(birddat1)[which(colnames(birddat1) == "PCODE_derived")] <- "PCODE"
colnames(birddat1)[which(colnames(birddat1) == "SS_derived")] <- "SS"

load("cache/offsets_all.RData")

birddat <- merge(vri_bird_tempaligned[c("PKEY")], birddat1[c("PKEY", "YYYY", "ROUND", "SPECIES", "BEH", "METHOD", "MaxDuration", "Maxdist", "ABUND")], by="PKEY", all.x=T)
birddat <- merge(birddat, offsets_all, by="PKEY", all.x=T)
```

### Subset to desired behaviours 

Following Peter Solymos's guidelines (https://github.com/psolymos/bamanalytics/blob/master/R/dataprocessing.R) and keeping behaviours 1, 6, and 11. See table below for definitions. 

``` {r}
t.behav <- data.frame(table(birddat$BEH))
colnames(t.behav) <- c("behcode", "NumObservations")

lookup.behav <- read.csv("data/LOOKUP.BehaviourCodes.csv", header=T)
lookup.behav <- lookup.behav[order(lookup.behav$behcode),]

t.behav <- merge(lookup.behav, t.behav, all=T)
t.behav$NumObservations[is.na(t.behav$NumObservations)] <- 0
kable(t.behav, row.names=F)

birddat2 <- birddat[birddat$BEH %in% c("1", "6", "11"),]
```

``` {r}
birddat3 <- birddat2[c("METHOD", "PKEY", "YYYY", "SPECIES", "BEH", "ABUND", "MaxDuration", "Maxdist", "ROUND", "Offset_simple", "Exposure_simple", "OFFSET")]
```

#### Reclass methods to the subset I would use as factors in subsequent analyses

``` {r, eval=T}
#distanceClass <- data.frame(Maxdist=unique(birddat3$Maxdist), distanceClass=c("unlimited", "80-100", "80-100", "50"))
distanceClass <- data.frame(Maxdist=unique(birddat3$Maxdist), distanceClass=c("unlimited", "80-100", "50"))

#durationClass <- data.frame(MaxDuration=unique(birddat3$MaxDuration), durationClass=c("3", "5-6", "8-10", "8-10", "5-6"))
durationClass <- data.frame(MaxDuration=unique(birddat3$MaxDuration), durationClass=c("5-6", "8-10", "5-6"))

birddat3.5 <- merge(birddat3, distanceClass, by="Maxdist", all.x=T)
birddat3.5 <- merge(birddat3.5, durationClass, by="MaxDuration", all.x=T)
colnames(birddat3.5)
rm(birddat3)
```

### Aggregate to "total bird" count per survey instance

##### Add English names in addition to  CODES

``` {r}
codes <- read.csv("data/EC_AVIAN_CORE_20150324.csv")
colnames(codes)[which(colnames(codes)=="Species_ID")] <- "SPECIES"

#merge species names into point count dataset
birddat4 <- merge(birddat3.5, codes, by="SPECIES", all.x=T)

#subset for necessary columns
birddat4 <- birddat4[c(colnames(birddat3.5), "English_Name")]
```

##### Cut any problematic English names

``` {r}
unique(birddat4[is.na(birddat4$English_Name),]$SPECIES)

birddat4 <- birddat4[birddat4$SPECIES != "NONE",]
```

#### Aggregate

``` {r}
sumABUND <- aggregate(birddat4$ABUND,  by=list(PKEY=birddat4$PKEY), FUN=sum)
colnames(sumABUND)[2] <- "sumABUND"

speciesCOUNT <- aggregate(birddat4$ABUND,  by=list(PKEY=birddat4$PKEY), FUN=length) 
colnames(speciesCOUNT)[2] <- "speciesCOUNT"

tmp <- merge(sumABUND, speciesCOUNT, by="PKEY")

birddat5 <- merge(tmp, birddat4[c("PKEY", "METHOD", "YYYY", "MaxDuration", "Maxdist", "durationClass", "distanceClass", "Exposure_simple", "Offset_simple", "OFFSET")], by="PKEY")
``` 

**Re-derive PCODE and SS from PKEY**

``` {r}
birddat5$PCODE_derived <- unlist(lapply(strsplit(as.character(birddat5$PKEY), ":", fixed=T), function(x) {x[1]}))
birddat5$SS_derived <- unlist(lapply(strsplit(as.character(birddat5$PKEY), ":", fixed=T), function(x) {paste(x[1:3], collapse=":")}))

birddat5 <- birddat5[!duplicated(birddat5$PKEY),]
```

### Combine bird data with VRI data

``` {r}
nrow(vri_bird_tempaligned)
length(unique(vri_bird_tempaligned$PKEY))

bird_vri_dat_ready <- merge(birddat5, vri_bird_tempaligned[-which(colnames(vri_bird_tempaligned) %in% c("FID_BIRD", "VRIintersectLayer", "OBJECTID", "FID_VRI", "BirdLayer", "MAP_ID", "POLYGON_ID", "OPENING_IND", "OPENING_SOURCE", "POLYGON_AREA", "INPUT_DATE", "INTERPRETATION_DATE", "REFERENCE_YEAR", "ATTRIBUTION_BASE_DATE", "PROJECTED_DATE", "OPENING_ID", "ORG_UNIT_NO", "ORG_UNIT_CODE", "FD_PERCENT", "PL_PERCENT", "AT_PERCENT", "VRI.layer.polygon.ID",  "ToleratedLag", "LagExceedsTolerated", "VRI.layer.polygon.ID.PKEY", "YYYY", "SS", "Offset", "Exposure"))], by="PKEY", all.x=T)
bird_vri_dat_ready$ForID <- with(bird_vri_dat_ready, paste(BecZone, SpeciesGroup, Age, Height, sep="."))
bird_vri_dat_ready <- bird_vri_dat_ready[sort(colnames(bird_vri_dat_ready))]

kable(bird_vri_dat_ready[sample(nrow(bird_vri_dat_ready), 10), c("ForID", "BecZone", "SpeciesGroup", "Age", "Height", "AgeClass_calc", "HeightClass_calc", "DisturbanceYear", "speciesCOUNT", "sumABUND")])
```

### Check if offsets are the same within a given survey method

* The following list should have only one value per list item if I'm using the simpler method of calculate offsets
    * It will have multiple values if I"m using the more complex approach
* Each list item represents a different method

``` {r}
methods <- unique(bird_vri_dat_ready$METHOD)

## function to look at unique offsets within a given method. Should have one value per row/method.
lapply(methods, function(x) {
  unique(bird_vri_dat_ready[bird_vri_dat_ready$METHOD == x,]$Offset_simple)
})

lapply(methods, function(x) {
  unique(bird_vri_dat_ready[bird_vri_dat_ready$METHOD == x,]$OFFSET)
})
```

``` {r}
cache("bird_vri_dat_ready")
```
